CmuxSwiftRender: split user-function evaluation out of ExpressionEvaluator
The dictionary-literal addition pushed ExpressionEvaluator.swift over its
file-length budget (651 vs 638). Move the value-returning user-function
section (parameter binding, block/if/switch value semantics) to
ExpressionEvaluator+UserFunctions.swift instead of refreshing the budget;
callValueFunction and evalBlockValue become internal for cross-file use
within the package. No behavior change; all 62 package tests pass.
